requirements.txt
groq
pydantic
PyPDF2
python-dotenv
python-docx
streamlit
requests
beautifulsoup4
schedule